notes
Arrow on side of page:
Add "MORE" text on last cell
Add "TOP" when you get to bottom

Footer
Logo on right needs to move in toward left, too close to the arrow on the sidebar.
Balance of the elements on the footer seems a little off.
Top align the first row of elements, Top align Maine stuff and Email stuff on the second row.

Scrolling issue
On this scroll, see how everything is blending together? Can we make it so the body text disappears behind an imaginary line so there isn't any overlap? It's really distracting to them and are asking that it's cleaned up.